chore: SelectPoolType as a separate component
