+++ apps/balancer-tools/src/app/poolsimulator/(components)/Menu.tsx
@@ -1,4 +1,3 @@
-import { Close, DialogClose } from "@radix-ui/react-dialog";
 import { MagnifyingGlassIcon } from "@radix-ui/react-icons";
 import * as Separator from "@radix-ui/react-separator";
 import { usePathname } from "next/navigation";
@@ -20,6 +19,7 @@ import {
 // import InitialDataForm from "./InitialDataForm"; BAL-499
 import { PoolParamsForm } from "./PoolParamsForm";
 import { SearchPoolFormDialog } from "./SearchPoolFormDialog";
+import { SelectPoolType } from "./SelectPoolType";
 
 const POOL_TYPES_MAPPER = {
   MetaStable: "Meta Stable",
@@ -104,40 +104,6 @@ function IndexMenu() {
         </div>
       </div>
     </SearchPoolFormWithDataForm>
-  );
-}
-
-function SelectPoolType({
-  selectedType,
-}: {
-  selectedType: PoolType;
-  setSelectedType: (value: PoolType) => void;
-}) {
-  const { setPoolType } = usePoolSimulator();
-
-  function onSubmit() {
-    setPoolType(selectedType);
-    Close;
-  }
-  return (
-    <div className="text-slate12 flex flex-col items-center gap-y-4">
-      <div className="text-center flex flex-col items-center">
-        <span>You are about to change the pool type to {selectedType}</span>
-        <span>This action will reset all the parameters.</span>
-      </div>
-      <div className="flex gap-x-4 py-3">
-        <DialogClose>
-          <span className="bg-transparent text-slate9 border-slate9 hover:bg-slate2 hover:border-slate2 rounded-md py-3 px-5 text-center text-sm font-semibold border focus-visible:outline-blue7 focus-visible:outline-offset-2 disabled:opacity-40">
-            Cancel
-          </span>
-        </DialogClose>
-        <DialogClose onClick={onSubmit}>
-          <span className="bg-blue9 text-slate12 hover:bg-blue10 border-blue9 rounded-md py-3 px-5 text-center text-sm font-semibold border focus-visible:outline-blue7 focus-visible:outline-offset-2 disabled:opacity-40">
-            Confirm
-          </span>
-        </DialogClose>
-      </div>
-    </div>
   );
 }
